requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
colorgram.py==1.2.0
idna==3.2
Pillow==8.3.2
python-dotenv==0.19.1
requests==2.26.0
urllib3==1.26.7
